requirements.txt
python-telegram-bot
aiosqlite~=0.19.0
prettytable~=3.9.0
Pillow~=10.1.0
telegram~=0.0.1